create useNowPlaying custom hook
